refactor: move unique implementation into helpers dir
